use f32 instead of f64

+++ raytracing/src/vec.rs
@@ -2,30 +2,26 @@ use core::ops;
 
 #[derive(Debug, Default, Clone, Copy, PartialEq)]
 pub struct Vec3 {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl Vec3 {
-    pub fn new(x: f64, y: f64, z: f64) -> Self {
+    pub fn new(x: f32, y: f32, z: f32) -> Self {
         Self { x, y, z }
     }
 
-    pub fn dot(&self, other: Vec3) -> f64 {
-        self.x * other.x +
-        self.y * other.y +
-        self.z * other.z
+    pub fn dot(&self, other: Vec3) -> f32 {
+        self.x * other.x + self.y * other.y + self.z * other.z
     }
 
-    pub fn len(&self) -> f64 {
+    pub fn len(&self) -> f32 {
         self.len2().sqrt()
     }
 
-    pub fn len2(&self) -> f64 {
-        self.x.powf(2.0) +
-        self.y.powf(2.0) +
-        self.z.powf(2.0)
+    pub fn len2(&self) -> f32 {
+        self.x.powf(2.0) + self.y.powf(2.0) + self.z.powf(2.0)
     }
 }
 
@@ -53,10 +49,10 @@ impl ops::Sub for Vec3 {
     }
 }
 
-impl ops::Mul<f64> for Vec3 {
+impl ops::Mul<f32> for Vec3 {
     type Output = Self;
 
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Self {
             x: self.x * rhs,
             y: self.y * rhs,
